refactor code, add workoutvalidator
