+++ package.json
@@ -19,6 +19,7 @@
     "express": "^4.18.2",
     "express-favicon": "^2.0.4",
     "express-rate-limit": "^7.5.0",
+    "express-validator": "^7.2.1",
     "helmet": "^8.1.0",
     "http-status-codes": "^2.3.0",
     "jsonwebtoken": "^9.0.2",
